website/st_pg_draw/requirements.txt
pygame-ce==2.4.0
pillow==10.0.1
numpy==1.26.1
streamlit_ace==0.1.1
streamlit==1.31.1